feat: 実行時間をfloat値に修正

+++ judge_server/src/judge_result.py
@@ -7,7 +7,7 @@ class JudgeResult():
             status: str,
             error: str = "",
             warning: str = "",
-            time: int = 0,
+            time: float = 0,
             memory: int = 0):
         self.status = status
         self.time = time
